backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI and ML dependencies
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.18
openai>=1.6.1,<2.0.0

# Data processing
pandas==2.1.4
numpy==1.24.3
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File handling and HTTP
aiofiles==23.2.1
httpx==0.25.2

# Database and caching
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Monitoring and logging
psutil==5.9.6
structlog==23.2.0

# Report generation
reportlab==4.0.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Production utilities
gunicorn==21.2.0
python-multipart==0.0.6
